refactor(test): Queries test suite to use ERPNextTestSuite
